requirements.txt
certifi==2020.4.5.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-import-export==2.2.0
django-nine==0.2.3
django-rest-elasticsearch==0.4.2
djangorestframework==3.11.0
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pep8==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
unicodecsv==0.14.1
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
